+++ src/NetCore/Web/Api/wwwroot/scripts/api/ui/webgis_uibuilder.js
@@ -719,7 +719,7 @@
                                 $(e).find('.webgis-ui-collapsable-title').trigger('click');
                             }
                         });
-                        $(this).find("img").attr('src', webgis.css.imgResource("expanded-26.png", "ui"));
+                        $(this).find("img.collapse-arrow").attr('src', webgis.css.imgResource("expanded-26.png", "ui"));
                         $content.slideDown();
 
                         $content
@@ -729,12 +729,13 @@
                             .removeClass('webgis-ui-collapsable-autoclick')
                     }
                     else {
-                        $(this).find("img").attr('src', webgis.css.imgResource("collapsed-26.png", "ui"));
+                        $(this).find("img.collapse-arrow").attr('src', webgis.css.imgResource("collapsed-26.png", "ui"));
                         $content.slideUp();
                     }
                 });
-            $("<h2><img src='" + webgis.css.imgResource(show ? "expanded-26.png" : "collapsed-26.png", "ui") + "'>" +
+            $("<h2><img src='" + webgis.css.imgResource(show ? "expanded-26.png" : "collapsed-26.png", "ui") + "' class='collapse-arrow' >" +
                 (element.type === 'optionscontainer' ? "&nbsp;<div class='webgis-ui-selected-option' style='display:inline-block;position:relative;top:7px;height:28px;overflow:hidden'></div>" : "") +
+                (element.icon ? "&nbsp;<img src='" + webgis.css.imgResource(element.icon) + "'/>" : "") +
                 "&nbsp;" + element.title + "</h2>").appendTo($collapsableTitle);
             $parent = $("<div>").addClass('webgis-ui-collapsable-content').css('display', show ? 'block' : 'none').appendTo($collapsable);
         }